+++ src/Firebase/V2/Http/Auth/AdminToken.php
@@ -2,9 +2,10 @@
 
 namespace Firebase\V2\Http\Auth;
 
-use Firebase\Token\TokenGenerator;
 use Firebase\V2\Http\Auth;
 use GuzzleHttp\Psr7;
+use Lcobucci\JWT\Builder;
+use Lcobucci\JWT\Signer\Hmac\Sha256;
 use Psr\Http\Message\RequestInterface;
 
 class AdminToken implements Auth
@@ -12,22 +13,27 @@ class AdminToken implements Auth
     /**
      * @var string
      */
-    private $token;
+    private $secret;
 
     public function __construct(string $databaseSecret)
     {
-        $options = ['admin' => true];
-
-        $this->token = (new TokenGenerator($databaseSecret))
-            ->setOptions($options)
-            ->create();
+        $this->secret = $databaseSecret;
     }
 
     public function authenticateRequest(RequestInterface $request): RequestInterface
     {
         $uri = $request->getUri();
 
-        $queryParams = ['auth' => $this->token] + Psr7\parse_query($uri->getQuery());
+        $now = time();
+
+        $token = (new Builder())
+            ->setIssuedAt($now)
+            ->setExpiration($now + (60 * 60))
+            ->set('admin', true)
+            ->sign(new Sha256(), $this->secret)
+            ->getToken();
+
+        $queryParams = ['auth' => (string) $token] + Psr7\parse_query($uri->getQuery());
         $queryString = Psr7\build_query($queryParams);
 
         $newUri = $uri->withQuery($queryString);